+++ src/test/java/bio/terra/service/filedata/google/firestore/FileOperationTest.java
@@ -1,14 +1,9 @@
 package bio.terra.service.filedata.google.firestore;
 
-import static bio.terra.common.PdaoConstant.PDAO_LOAD_HISTORY_TABLE;
 import static org.hamcrest.CoreMatchers.containsString;
 import static org.hamcrest.CoreMatchers.equalTo;
 import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.Matchers.greaterThan;
-import static org.junit.Assert.assertNotNull;
-import static org.junit.Assert.assertNull;
 import static org.junit.Assert.assertTrue;
-import static org.junit.Assert.fail;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.doReturn;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
@@ -19,15 +14,9 @@ import bio.terra.common.EmbeddedDatabaseTest;
 import bio.terra.common.TestUtils;
 import bio.terra.common.category.Connected;
 import bio.terra.common.fixtures.ConnectedOperations;
-import bio.terra.common.fixtures.JsonLoader;
 import bio.terra.common.fixtures.Names;
 import bio.terra.model.BillingProfileModel;
-import bio.terra.model.BulkLoadArrayRequestModel;
-import bio.terra.model.BulkLoadArrayResultModel;
 import bio.terra.model.BulkLoadFileModel;
-import bio.terra.model.BulkLoadFileResultModel;
-import bio.terra.model.BulkLoadFileState;
-import bio.terra.model.BulkLoadRequestModel;
 import bio.terra.model.BulkLoadResultModel;
 import bio.terra.model.ConfigGroupModel;
 import bio.terra.model.ConfigModel;
@@ -40,28 +29,11 @@ import bio.terra.service.configuration.ConfigEnum;
 import bio.terra.service.configuration.ConfigurationService;
 import bio.terra.service.dataset.DatasetDao;
 import bio.terra.service.dataset.DatasetDaoUtils;
-import bio.terra.service.filedata.DrsIdService;
-import bio.terra.service.filedata.google.gcs.GcsChannelWriter;
 import bio.terra.service.iam.IamProviderInterface;
 import bio.terra.service.resourcemanagement.ResourceService;
 import bio.terra.service.resourcemanagement.google.GoogleProjectService;
-import bio.terra.service.resourcemanagement.google.GoogleResourceConfiguration;
-import bio.terra.service.tabulardata.google.BigQueryPdao;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.google.cloud.bigquery.FieldValueList;
-import com.google.cloud.bigquery.TableResult;
-import com.google.cloud.storage.Storage;
-import com.google.cloud.storage.StorageOptions;
 import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
-import java.io.IOException;
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.Set;
 import java.util.UUID;
-import java.util.stream.Collectors;
-import org.hamcrest.Matchers;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Ignore;
@@ -90,17 +62,12 @@ import org.springframework.test.web.servlet.MvcResult;
 @Category(Connected.class)
 @EmbeddedDatabaseTest
 public class FileOperationTest {
+
   @Autowired private MockMvc mvc;
-  @Autowired private ObjectMapper objectMapper;
-  @Autowired private JsonLoader jsonLoader;
   @Autowired private ConnectedTestConfiguration testConfig;
-  @Autowired private DrsIdService drsService;
-  @Autowired private GoogleResourceConfiguration googleResourceConfiguration;
   @Autowired private ConnectedOperations connectedOperations;
   @Autowired private ConfigurationService configService;
   @Autowired private DatasetDao datasetDao;
-  @Autowired private ResourceService dataLocationService;
-  @Autowired private BigQueryPdao bigQueryPdao;
 
   @MockBean private IamProviderInterface samService;
 
@@ -319,327 +286,7 @@ public class FileOperationTest {
         datasetDao);
   }
 
-  // -- array bulk load --
-
-  @Test
-  public void arrayMultiFileLoadSuccessTest() throws Exception {
-    int fileCount = 10;
-
-    // Test copying load_history data in chunks
-    ConfigModel loadHistoryChunkSize =
-        configService.getConfig(ConfigEnum.LOAD_HISTORY_COPY_CHUNK_SIZE.name());
-    loadHistoryChunkSize.setParameter(new ConfigParameterModel().value("2"));
-    ConfigGroupModel configGroupModel =
-        new ConfigGroupModel()
-            .label("FileOperationTestMultiFileLoad")
-            .addGroupItem(loadHistoryChunkSize);
-    configService.setConfig(configGroupModel);
-
-    BulkLoadArrayRequestModel arrayLoad =
-        makeSuccessArrayLoad("arrayMultiFileLoadSuccessTest", 0, fileCount);
-
-    BulkLoadArrayResultModel result =
-        connectedOperations.ingestArraySuccess(datasetSummary.getId(), arrayLoad);
-    checkLoadSummary(result.getLoadSummary(), arrayLoad.getLoadTag(), fileCount, fileCount, 0, 0);
-
-    var loadHistoryList =
-        connectedOperations.getLoadHistory(
-            datasetSummary.getId(), result.getLoadSummary().getLoadTag(), 0, 20);
-
-    assertThat(
-        "The correct number of load history entries are returned",
-        loadHistoryList.getTotal(),
-        equalTo(fileCount));
-
-    assertThat(
-        "getting load history has the same items as response from bulk file load",
-        loadHistoryList.getItems().stream()
-            .map(TestUtils::toBulkLoadFileResultModel)
-            .collect(Collectors.toSet()),
-        Matchers.equalTo(Set.copyOf(result.getLoadFileResults())));
-
-    Map<String, String> fileIdMap = new HashMap<>();
-    for (BulkLoadFileResultModel fileResult : result.getLoadFileResults()) {
-      checkFileResultSuccess(fileResult);
-      fileIdMap.put(fileResult.getTargetPath(), fileResult.getFileId());
-    }
-
-    // Query Big Query datarepo_load_history table - should reflect all files loaded above
-    String columnToQuery = "file_id";
-    TableResult queryLoadHistoryTableResult = queryLoadHistoryTable(columnToQuery);
-    ArrayList<String> ids = new ArrayList<>();
-    queryLoadHistoryTableResult
-        .iterateAll()
-        .forEach(r -> ids.add(r.get(columnToQuery).getStringValue()));
-
-    assertThat(
-        "Number of files in datarepo_load_history table match load summary",
-        ids.size(),
-        equalTo(fileCount));
-    for (String bq_file_id : ids) {
-      assertNotNull(
-          "fileIdMap should contain File_id from datarepo_load_history",
-          fileIdMap.containsValue(bq_file_id));
-    }
-
-    // retry successful load to make sure it still succeeds and does nothing
-    BulkLoadArrayResultModel result2 =
-        connectedOperations.ingestArraySuccess(datasetSummary.getId(), arrayLoad);
-    checkLoadSummary(result2.getLoadSummary(), arrayLoad.getLoadTag(), fileCount, fileCount, 0, 0);
-
-    for (BulkLoadFileResultModel fileResult : result.getLoadFileResults()) {
-      checkFileResultSuccess(fileResult);
-      assertThat(
-          "FileId matches",
-          fileResult.getFileId(),
-          equalTo(fileIdMap.get(fileResult.getTargetPath())));
-    }
-  }
-
-  // Get the count of rows in a table or view
-  private TableResult queryLoadHistoryTable(String columns) throws Exception {
-    return TestUtils.selectFromBigQueryDataset(
-        bigQueryPdao,
-        datasetDao,
-        dataLocationService,
-        datasetSummary.getName(),
-        PDAO_LOAD_HISTORY_TABLE,
-        columns);
-  }
-
-  @Test
-  public void arrayMultiFileLoadDoubleSuccessTest() throws Exception {
-    int fileCount = 8;
-    int totalfileCount = fileCount * 2;
-    BulkLoadArrayRequestModel arrayLoad1 =
-        makeSuccessArrayLoad("arrayMultiDoubleSuccess", 0, fileCount);
-    BulkLoadArrayRequestModel arrayLoad2 =
-        makeSuccessArrayLoad("arrayMultiDoubleSuccess", fileCount, fileCount);
-    String loadTag1 = arrayLoad1.getLoadTag();
-    String loadTag2 = arrayLoad2.getLoadTag();
-    UUID datasetId = datasetSummary.getId();
-
-    MvcResult result1 = connectedOperations.ingestArrayRaw(datasetId, arrayLoad1);
-    MvcResult result2 = connectedOperations.ingestArrayRaw(datasetId, arrayLoad2);
-
-    MockHttpServletResponse response1 = connectedOperations.validateJobModelAndWait(result1);
-    MockHttpServletResponse response2 = connectedOperations.validateJobModelAndWait(result2);
-
-    BulkLoadArrayResultModel resultModel1 =
-        connectedOperations.handleSuccessCase(response1, BulkLoadArrayResultModel.class);
-
-    BulkLoadArrayResultModel resultModel2 =
-        connectedOperations.handleSuccessCase(response2, BulkLoadArrayResultModel.class);
-
-    checkLoadSummary(resultModel1.getLoadSummary(), loadTag1, fileCount, fileCount, 0, 0);
-    checkLoadSummary(resultModel2.getLoadSummary(), loadTag2, fileCount, fileCount, 0, 0);
-
-    List<String> fileIds = new ArrayList<>();
-    for (BulkLoadFileResultModel fileResult : resultModel1.getLoadFileResults()) {
-      checkFileResultSuccess(fileResult);
-      fileIds.add(fileResult.getFileId());
-    }
-    for (BulkLoadFileResultModel fileResult : resultModel2.getLoadFileResults()) {
-      checkFileResultSuccess(fileResult);
-      fileIds.add(fileResult.getFileId());
-    }
-
-    // Query Big Query datarepo_load_history table - should reflect all files loaded above
-    String columnToQuery = "file_id";
-    TableResult queryLoadHistoryTableResult = queryLoadHistoryTable(columnToQuery);
-    ArrayList<String> bq_fileIds = new ArrayList<>();
-    queryLoadHistoryTableResult
-        .iterateAll()
-        .forEach(r -> bq_fileIds.add(r.get(columnToQuery).getStringValue()));
-
-    assertThat(
-        "Number of files in datarepo_load_history table match load summary",
-        totalfileCount,
-        equalTo(bq_fileIds.size()));
-    for (String bq_file_id : bq_fileIds) {
-      assertNotNull(
-          "fileIdMap should contain File_id from datarepo_load_history",
-          fileIds.contains(bq_file_id));
-    }
-  }
-
-  @Test
-  public void arrayMultiFileLoadFailRetryTest() throws Exception {
-    String testId = Names.randomizeName("test");
-    String loadTag = "arrayMultiFileLoadFileRetryTest" + testId;
-    BulkLoadArrayRequestModel arrayLoad =
-        new BulkLoadArrayRequestModel()
-            .profileId(profileModel.getId())
-            .loadTag(loadTag)
-            .maxFailedFileLoads(2);
-    arrayLoad.addLoadArrayItem(getFileModel(true, 2, testId));
-    arrayLoad.addLoadArrayItem(getFileModel(false, 3, testId));
-    arrayLoad.addLoadArrayItem(getFileModel(true, 4, testId));
-
-    BulkLoadArrayResultModel result =
-        connectedOperations.ingestArraySuccess(datasetSummary.getId(), arrayLoad);
-    checkLoadSummary(result.getLoadSummary(), loadTag, 3, 2, 1, 0);
-
-    Map<String, BulkLoadFileResultModel> resultMap = new HashMap<>();
-    for (BulkLoadFileResultModel fileResult : result.getLoadFileResults()) {
-      resultMap.put(fileResult.getTargetPath(), fileResult);
-    }
-    // Query Big Query datarepo_load_history table - assert correctly reflects different
-    // bulk load file states
-    String columnsToQuery = "state, file_id, error";
-    TableResult queryLoadHistoryTableResult = queryLoadHistoryTable(columnsToQuery);
-    for (FieldValueList item : queryLoadHistoryTableResult.getValues()) {
-      String state = item.get(0).getStringValue();
-      assertTrue(
-          "state should either be succeeded or failed.",
-          state.equals(BulkLoadFileState.SUCCEEDED.toString())
-              || state.equals(BulkLoadFileState.FAILED.toString()));
-      if (state.equals(BulkLoadFileState.SUCCEEDED.toString())) {
-        assertTrue("file_id should have value", item.get(1).getStringValue().length() > 0);
-        assertTrue("Error column should be empty", item.get(2).getStringValue().length() == 0);
-      } else if (state.equals(BulkLoadFileState.FAILED.toString())) {
-        assertTrue("file_id should NOT have value", item.get(1).getStringValue().length() == 0);
-        assertTrue("Error column should have value", item.get(2).getStringValue().length() > 0);
-      }
-    }
-    FieldValueList curr_result;
-
-    List<BulkLoadFileModel> loadArray = arrayLoad.getLoadArray();
-    BulkLoadFileResultModel fileResult = resultMap.get(loadArray.get(0).getTargetPath());
-    checkFileResultSuccess(fileResult);
-
-    fileResult = resultMap.get(loadArray.get(1).getTargetPath());
-    checkFileResultFailed(fileResult);
-
-    fileResult = resultMap.get(loadArray.get(2).getTargetPath());
-    checkFileResultSuccess(fileResult);
-
-    // fix the bad file and retry load
-    loadArray.set(1, getFileModel(true, 3, testId));
-    BulkLoadArrayResultModel result2 =
-        connectedOperations.ingestArraySuccess(datasetSummary.getId(), arrayLoad);
-    checkLoadSummary(result2.getLoadSummary(), loadTag, 3, 3, 0, 0);
-  }
-
-  @Test
-  public void arrayMultiFileLoadExceedMaxTest() throws Exception {
-    // Set the allowed array files to a very small value so we can easily hit the error
-    ConfigModel bulkArrayMaxConfig =
-        configService.getConfig(ConfigEnum.LOAD_BULK_ARRAY_FILES_MAX.name());
-    bulkArrayMaxConfig.setParameter(new ConfigParameterModel().value("5"));
-    ConfigGroupModel configGroupModel =
-        new ConfigGroupModel()
-            .label("FileOperationTest:loadExceedMax")
-            .addGroupItem(bulkArrayMaxConfig);
-    configService.setConfig(configGroupModel);
-
-    String testId = Names.randomizeName("test");
-    String loadTag = "arrayMultiFileLoadExceedMaxTest" + testId;
-    BulkLoadArrayRequestModel arrayLoad =
-        new BulkLoadArrayRequestModel()
-            .profileId(profileModel.getId())
-            .loadTag(loadTag)
-            .maxFailedFileLoads(0);
-    for (int i = 0; i < 8; i++) {
-      arrayLoad.addLoadArrayItem(getFileModel(true, i, testId));
-    }
-
-    MvcResult result = connectedOperations.ingestArrayRaw(datasetSummary.getId(), arrayLoad);
-    assertThat(
-        "Got bad request",
-        result.getResponse().getStatus(),
-        equalTo(HttpStatus.BAD_REQUEST.value()));
-  }
-
-  // -- file bulk load --
-  @Test
-  public void multiFileLoadSuccessTest() throws Exception {
-    BulkLoadRequestModel loadRequest =
-        makeBulkFileLoad(
-            "multiFileLoadSuccessTest", 0, 0, false, new boolean[] {true, true, true, true});
-
-    BulkLoadResultModel result =
-        connectedOperations.ingestBulkFileSuccess(datasetSummary.getId(), loadRequest);
-    checkLoadSummary(result, loadRequest.getLoadTag(), 4, 4, 0, 0);
-
-    // retry successful load to make sure it still succeeds and does nothing
-    result = connectedOperations.ingestBulkFileSuccess(datasetSummary.getId(), loadRequest);
-    checkLoadSummary(result, loadRequest.getLoadTag(), 4, 4, 0, 0);
-  }
-
-  @Test
-  public void multiFileLoadFailRetryTest() throws Exception {
-    BulkLoadRequestModel loadRequest =
-        makeBulkFileLoad(
-            "multiFileLoadFailRetry", 0, 0, false, new boolean[] {true, false, true, false});
-    loadRequest.maxFailedFileLoads(4);
-    String loadTag = loadRequest.getLoadTag();
-
-    BulkLoadResultModel result =
-        connectedOperations.ingestBulkFileSuccess(datasetSummary.getId(), loadRequest);
-    checkLoadSummary(result, loadTag, 4, 2, 2, 0);
-
-    loadRequest =
-        makeBulkFileLoad(
-            "multiFileLoadFailRetry", 0, 0, false, new boolean[] {true, true, true, true});
-    loadRequest.loadTag(loadTag);
-    result = connectedOperations.ingestBulkFileSuccess(datasetSummary.getId(), loadRequest);
-    checkLoadSummary(result, loadTag, 4, 4, 0, 0);
-  }
-
-  @Test
-  public void multiFileLoadSuccessExtraKeysTest() throws Exception {
-    BulkLoadRequestModel loadRequest =
-        makeBulkFileLoad(
-            "multiFileLoadSuccessExtraKeys", 0, 0, true, new boolean[] {true, true, true, true});
-
-    BulkLoadResultModel result =
-        connectedOperations.ingestBulkFileSuccess(datasetSummary.getId(), loadRequest);
-    checkLoadSummary(result, loadRequest.getLoadTag(), 4, 4, 0, 0);
-  }
-
-  @Test
-  public void multiFileLoadBadLineTest() throws Exception {
-    // part 1: test that we exit with the bad line error when we have fewer than the max
-    BulkLoadRequestModel loadRequest =
-        makeBulkFileLoad(
-            "multiFileLoadBadLineSuccess", 0, 3, false, new boolean[] {true, false, true, false});
-    loadRequest.maxFailedFileLoads(4);
-
-    ErrorModel errorModel =
-        connectedOperations.ingestBulkFileFailure(datasetSummary.getId(), loadRequest);
-    assertThat(
-        "Expected error", errorModel.getMessage(), containsString("bad lines in the control file"));
-    assertThat("Expected error", errorModel.getMessage(), containsString("There were"));
-    assertThat("Expected number of error details", errorModel.getErrorDetail().size(), equalTo(3));
-
-    // part 2: test that we exit with bad line error when we have more than the max
-    loadRequest =
-        makeBulkFileLoad(
-            "multiFileLoadBadLineSuccess", 0, 6, false, new boolean[] {true, true, true, true});
-    loadRequest.maxFailedFileLoads(4);
-
-    errorModel = connectedOperations.ingestBulkFileFailure(datasetSummary.getId(), loadRequest);
-    assertThat(
-        "Expected error", errorModel.getMessage(), containsString("bad lines in the control file"));
-    assertThat("Expected error", errorModel.getMessage(), containsString("More than"));
-    assertThat(
-        "Expected number of error details", errorModel.getErrorDetail().size(), greaterThan(5));
-  }
-
-  private static void checkFileResultFailed(BulkLoadFileResultModel fileResult) {
-    assertNotNull("Error is not null", fileResult.getError());
-    assertNull("FileId is null", fileResult.getFileId());
-    assertThat("State is FAILED", fileResult.getState(), equalTo(BulkLoadFileState.FAILED));
-  }
-
-  private static void checkFileResultSuccess(BulkLoadFileResultModel fileResult) {
-    assertNull("Error is null", fileResult.getError());
-    assertNotNull("FileId is not null", fileResult.getFileId());
-    assertThat("State is SUCCEEDED", fileResult.getState(), equalTo(BulkLoadFileState.SUCCEEDED));
-  }
-
-  private static void checkLoadSummary(
+  static void checkLoadSummary(
       BulkLoadResultModel summary,
       String loadTag,
       int total,
@@ -651,67 +298,6 @@ public class FileOperationTest {
     assertThat("correct succeeded files", summary.getSucceededFiles(), equalTo(succeeded));
     assertThat("correct failed files", summary.getFailedFiles(), equalTo(failed));
     assertThat("correct notTried files", summary.getNotTriedFiles(), equalTo(notTried));
-  }
-
-  private BulkLoadArrayRequestModel makeSuccessArrayLoad(
-      String tagBase, int startIndex, int fileCount) {
-    String testId = Names.randomizeName("test");
-    String loadTag = tagBase + testId;
-    BulkLoadArrayRequestModel arrayLoad =
-        new BulkLoadArrayRequestModel()
-            .profileId(profileModel.getId())
-            .loadTag(loadTag)
-            .maxFailedFileLoads(0);
-    for (int index = startIndex; index < startIndex + fileCount; index++) {
-      arrayLoad.addLoadArrayItem(getFileModel(true, index, testId));
-    }
-    return arrayLoad;
-  }
-
-  private BulkLoadRequestModel makeBulkFileLoad(
-      String tagBase, int startIndex, int badLines, boolean addExtraKeys, boolean[] validPattern) {
-    int fileCount = validPattern.length;
-    String testId = Names.randomizeName("test");
-    String loadTag = tagBase + testId;
-    String targetPath = "scratch/controlfile" + UUID.randomUUID().toString() + ".json";
-    connectedOperations.addScratchFile(targetPath); // track the file so it gets cleaned up
-
-    String gspath = "gs://" + testConfig.getIngestbucket() + "/" + targetPath;
-    Storage storage = StorageOptions.getDefaultInstance().getService();
-
-    try (GcsChannelWriter writer =
-        new GcsChannelWriter(storage, testConfig.getIngestbucket(), targetPath)) {
-      for (int i = 0; i < badLines; i++) {
-        String badLine = "bad line: " + loadTag + "\n";
-        writer.write(badLine);
-      }
-
-      for (int i = 0; i < fileCount; i++) {
-        BulkLoadFileModel fileModel = getFileModel(validPattern[i], startIndex + i, testId);
-        String fileLine = objectMapper.writeValueAsString(fileModel) + "\n";
-        // Inject extra key-value pairs into file lines
-        if (addExtraKeys) {
-          fileLine = fileLine.replaceFirst("^\\{", "{\"customKey\":\"customValue\",");
-          logger.info("Added extra keys: " + fileLine);
-        }
-        writer.write(fileLine);
-      }
-    } catch (IOException ex) {
-      fail(
-          "Failed to write load file '"
-              + targetPath
-              + "' to bucket '"
-              + testConfig.getIngestbucket()
-              + "'");
-    }
-
-    BulkLoadRequestModel loadRequest =
-        new BulkLoadRequestModel()
-            .profileId(profileModel.getId())
-            .loadTag(loadTag)
-            .maxFailedFileLoads(0)
-            .loadControlFile(gspath);
-    return loadRequest;
   }
 
   // We have a static array of good paths and bad paths with their associated
@@ -781,7 +367,7 @@ public class FileOperationTest {
         "/encodefiles/20180504/ENCFF823AJQ.bam.bai"
       };
 
-  private BulkLoadFileModel getFileModel(boolean getGood, int index, String testId) {
+  static BulkLoadFileModel getFileModel(boolean getGood, int index, String testId) {
     assertTrue("test bug: file index not in range", index < fileTarget.length);
 
     BulkLoadFileModel model = new BulkLoadFileModel().mimeType("application/binary");
