+++ src/test/java/bio/terra/service/snapshot/SnapshotConnectedTest.java
@@ -14,7 +14,6 @@ import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.when;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.delete;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
 
 import bio.terra.app.configuration.ConnectedTestConfiguration;
@@ -23,27 +22,20 @@ import bio.terra.common.TestUtils;
 import bio.terra.common.category.Connected;
 import bio.terra.common.fixtures.ConnectedOperations;
 import bio.terra.common.fixtures.JsonLoader;
-import bio.terra.common.fixtures.Names;
 import bio.terra.model.BillingProfileModel;
-import bio.terra.model.DRSObject;
 import bio.terra.model.DatasetSummaryModel;
 import bio.terra.model.DeleteResponseModel;
 import bio.terra.model.EnumerateSnapshotModel;
 import bio.terra.model.ErrorModel;
-import bio.terra.model.FileLoadModel;
-import bio.terra.model.FileModel;
 import bio.terra.model.IngestRequestModel;
 import bio.terra.model.RelationshipModel;
 import bio.terra.model.SnapshotModel;
-import bio.terra.model.SnapshotRequestContentsModel;
 import bio.terra.model.SnapshotRequestModel;
-import bio.terra.model.SnapshotSourceModel;
 import bio.terra.model.SnapshotSummaryModel;
 import bio.terra.model.TableModel;
 import bio.terra.service.configuration.ConfigEnum;
 import bio.terra.service.configuration.ConfigurationService;
 import bio.terra.service.dataset.DatasetDao;
-import bio.terra.service.filedata.DrsId;
 import bio.terra.service.filedata.DrsIdService;
 import bio.terra.service.iam.IamProviderInterface;
 import bio.terra.service.iam.IamRole;
@@ -54,23 +46,17 @@ import com.google.cloud.bigquery.FieldValue;
 import com.google.cloud.bigquery.FieldValueList;
 import com.google.cloud.bigquery.QueryJobConfiguration;
 import com.google.cloud.bigquery.TableResult;
-import com.google.cloud.storage.BlobInfo;
 import com.google.cloud.storage.Storage;
 import com.google.cloud.storage.StorageOptions;
-import java.net.URI;
-import java.nio.charset.StandardCharsets;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.Set;
 import java.util.UUID;
-import java.util.concurrent.TimeUnit;
 import java.util.function.Function;
 import java.util.stream.Collectors;
-import org.apache.commons.io.IOUtils;
 import org.apache.commons.lang3.StringUtils;
-import org.hamcrest.CoreMatchers;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Ignore;
@@ -117,7 +103,7 @@ public class SnapshotConnectedTest {
 
   private String snapshotOriginalName;
   private BillingProfileModel billingProfile;
-  private Storage storage = StorageOptions.getDefaultInstance().getService();
+  private final Storage storage = StorageOptions.getDefaultInstance().getService();
   private DatasetSummaryModel datasetSummary;
 
   @Before
@@ -126,43 +112,23 @@ public class SnapshotConnectedTest {
     configService.reset();
     billingProfile =
         connectedOperations.createProfileForAccount(testConfig.getGoogleBillingAccountId());
-    datasetSummary = createTestDataset("snapshot-test-dataset.json");
-    loadCsvData(datasetSummary.getId(), "thetable", "snapshot-test-dataset-data-row-ids.csv");
+    datasetSummary =
+        SnapshotConnectedTestUtils.createTestDataset(
+            connectedOperations, billingProfile, "snapshot-test-dataset.json");
+    SnapshotConnectedTestUtils.loadCsvData(
+        connectedOperations,
+        jsonLoader,
+        storage,
+        testConfig.getIngestbucket(),
+        datasetSummary.getId(),
+        "thetable",
+        "snapshot-test-dataset-data-row-ids.csv");
   }
 
   @After
   public void tearDown() throws Exception {
     connectedOperations.teardown();
     configService.reset();
-  }
-
-  @Test
-  public void testHappyPath() throws Exception {
-    snapshotHappyPathTestingHelper("snapshot-test-snapshot.json");
-  }
-
-  @Test
-  public void testFaultyPath() throws Exception {
-    // Run the happy path test, but insert the GRANT ACCESS faults to simulate IAM propagation
-    // failures
-    configService.setFault(ConfigEnum.DATASET_GRANT_ACCESS_FAULT.name(), true);
-    configService.setFault(ConfigEnum.SNAPSHOT_GRANT_ACCESS_FAULT.name(), true);
-    testHappyPath();
-  }
-
-  @Test
-  public void testRowIdsHappyPath() throws Exception {
-    snapshotHappyPathTestingHelper("snapshot-row-ids-test-snapshot.json");
-  }
-
-  @Test
-  public void testQueryHappyPath() throws Exception {
-    snapshotHappyPathTestingHelper("snapshot-query-test-snapshot.json");
-  }
-
-  @Test
-  public void testFullViewsHappyPath() throws Exception {
-    snapshotHappyPathTestingHelper("snapshot-fullviews-test-snapshot.json");
   }
 
   @Test
@@ -177,11 +143,13 @@ public class SnapshotConnectedTest {
     assertThat("dataset samples loaded properly", datasetSamples, equalTo(5L));
 
     SnapshotRequestModel snapshotRequest =
-        makeSnapshotTestRequest(datasetMinimalSummary, "dataset-minimal-snapshot.json");
+        SnapshotConnectedTestUtils.makeSnapshotTestRequest(
+            jsonLoader, datasetMinimalSummary, "dataset-minimal-snapshot.json");
     MockHttpServletResponse response = performCreateSnapshot(snapshotRequest, "");
     SnapshotSummaryModel summaryModel = validateSnapshotCreated(snapshotRequest, response);
     SnapshotModel snapshotModel =
-        getTestSnapshot(summaryModel.getId(), snapshotRequest, datasetMinimalSummary);
+        SnapshotConnectedTestUtils.getTestSnapshot(
+            mvc, objectMapper, summaryModel.getId(), snapshotRequest, datasetMinimalSummary);
     List<TableModel> tables = snapshotModel.getTables();
     Optional<TableModel> participantTable =
         tables.stream().filter(t -> t.getName().equals("participant")).findFirst();
@@ -234,13 +202,15 @@ public class SnapshotConnectedTest {
 
     // TODO put big snapshot request into a GCS bucket
     SnapshotRequestModel snapshotRequestScale =
-        makeSnapshotTestRequest(datasetSummary, "hca-mvp-analysis-file-row-ids-snapshot.json");
+        SnapshotConnectedTestUtils.makeSnapshotTestRequest(
+            jsonLoader, datasetSummary, "hca-mvp-analysis-file-row-ids-snapshot.json");
 
     MockHttpServletResponse response = performCreateSnapshot(snapshotRequestScale, "");
     SnapshotSummaryModel summaryModel = validateSnapshotCreated(snapshotRequestScale, response);
 
     SnapshotModel snapshotModel =
-        getTestSnapshot(summaryModel.getId(), snapshotRequestScale, datasetSummary);
+        SnapshotConnectedTestUtils.getTestSnapshot(
+            mvc, objectMapper, summaryModel.getId(), snapshotRequestScale, datasetSummary);
 
     connectedOperations.deleteTestSnapshot(snapshotModel.getId());
     // Duplicate delete should work
@@ -260,10 +230,12 @@ public class SnapshotConnectedTest {
     assertThat("dataset samples loaded properly", datasetSamples, equalTo(5L));
 
     SnapshotRequestModel snapshotRequest =
-        makeSnapshotTestRequest(datasetArraySummary, "snapshot-array-struct.json");
+        SnapshotConnectedTestUtils.makeSnapshotTestRequest(
+            jsonLoader, datasetArraySummary, "snapshot-array-struct.json");
     MockHttpServletResponse response = performCreateSnapshot(snapshotRequest, "");
     SnapshotSummaryModel summaryModel = validateSnapshotCreated(snapshotRequest, response);
-    getTestSnapshot(summaryModel.getId(), snapshotRequest, datasetArraySummary);
+    SnapshotConnectedTestUtils.getTestSnapshot(
+        mvc, objectMapper, summaryModel.getId(), snapshotRequest, datasetArraySummary);
 
     BigQueryProject bigQuerySnaphsotProject =
         TestUtils.bigQueryProjectForSnapshotName(snapshotDao, summaryModel.getName());
@@ -278,8 +250,9 @@ public class SnapshotConnectedTest {
   public void testMinimalBadAsset() throws Exception {
     DatasetSummaryModel datasetMinimalSummary = setupMinimalDataset();
     SnapshotRequestModel snapshotRequest =
-        makeSnapshotTestRequest(datasetMinimalSummary, "dataset-minimal-snapshot-bad-asset.json");
-    MvcResult result = launchCreateSnapshot(snapshotRequest, "");
+        SnapshotConnectedTestUtils.makeSnapshotTestRequest(
+            jsonLoader, datasetMinimalSummary, "dataset-minimal-snapshot-bad-asset.json");
+    MvcResult result = SnapshotConnectedTestUtils.launchCreateSnapshot(mvc, snapshotRequest, "");
     MockHttpServletResponse response = connectedOperations.validateJobModelAndWait(result);
     assertThat(response.getStatus(), equalTo(HttpStatus.NOT_FOUND.value()));
   }
@@ -287,7 +260,8 @@ public class SnapshotConnectedTest {
   @Test
   public void testEnumeration() throws Exception {
     SnapshotRequestModel snapshotRequest =
-        makeSnapshotTestRequest(datasetSummary, "snapshot-test-snapshot.json");
+        SnapshotConnectedTestUtils.makeSnapshotTestRequest(
+            jsonLoader, datasetSummary, "snapshot-test-snapshot.json");
 
     // Other unit tests exercise the array bounds, so here we don't fuss with that here.
     // Just make sure we get the same snapshot summary that we made.
@@ -332,7 +306,8 @@ public class SnapshotConnectedTest {
   @Test
   public void testBadData() throws Exception {
     SnapshotRequestModel badDataRequest =
-        makeSnapshotTestRequest(datasetSummary, "snapshot-test-snapshot-baddata.json");
+        SnapshotConnectedTestUtils.makeSnapshotTestRequest(
+            jsonLoader, datasetSummary, "snapshot-test-snapshot-baddata.json");
 
     MockHttpServletResponse response = performCreateSnapshot(badDataRequest, "_baddata_");
     ErrorModel errorModel = handleCreateSnapshotFailureCase(response);
@@ -343,13 +318,15 @@ public class SnapshotConnectedTest {
   public void testDuplicateName() throws Exception {
     // create a snapshot
     SnapshotRequestModel snapshotRequest =
-        makeSnapshotTestRequest(datasetSummary, "snapshot-test-snapshot.json");
+        SnapshotConnectedTestUtils.makeSnapshotTestRequest(
+            jsonLoader, datasetSummary, "snapshot-test-snapshot.json");
     MockHttpServletResponse response = performCreateSnapshot(snapshotRequest, "_dup_");
     SnapshotSummaryModel summaryModel = validateSnapshotCreated(snapshotRequest, response);
 
     // fetch the snapshot and confirm the metadata matches the request
     SnapshotModel snapshotModel =
-        getTestSnapshot(summaryModel.getId(), snapshotRequest, datasetSummary);
+        SnapshotConnectedTestUtils.getTestSnapshot(
+            mvc, objectMapper, summaryModel.getId(), snapshotRequest, datasetSummary);
     assertNotNull("fetched snapshot successfully after creation", snapshotModel);
 
     // check that the snapshot metadata row is unlocked
@@ -368,7 +345,8 @@ public class SnapshotConnectedTest {
 
     // fetch the snapshot and confirm the metadata still matches the original
     SnapshotModel origModel =
-        getTestSnapshot(summaryModel.getId(), snapshotRequest, datasetSummary);
+        SnapshotConnectedTestUtils.getTestSnapshot(
+            mvc, objectMapper, summaryModel.getId(), snapshotRequest, datasetSummary);
     assertEquals("fetched snapshot remains unchanged", snapshotModel, origModel);
 
     // delete and confirm deleted
@@ -379,18 +357,29 @@ public class SnapshotConnectedTest {
   @Test
   public void testDeleteRecreateSnapshot() throws Exception {
     // create a dataset and load some tabular data
-    DatasetSummaryModel datasetSummary = createTestDataset("snapshot-test-dataset.json");
-    loadCsvData(datasetSummary.getId(), "thetable", "snapshot-test-dataset-data-row-ids.csv");
+    DatasetSummaryModel datasetSummary =
+        SnapshotConnectedTestUtils.createTestDataset(
+            connectedOperations, billingProfile, "snapshot-test-dataset.json");
+    SnapshotConnectedTestUtils.loadCsvData(
+        connectedOperations,
+        jsonLoader,
+        storage,
+        testConfig.getIngestbucket(),
+        datasetSummary.getId(),
+        "thetable",
+        "snapshot-test-dataset-data-row-ids.csv");
 
     // create a snapshot
     SnapshotRequestModel snapshotRequest =
-        makeSnapshotTestRequest(datasetSummary, "snapshot-test-snapshot.json");
+        SnapshotConnectedTestUtils.makeSnapshotTestRequest(
+            jsonLoader, datasetSummary, "snapshot-test-snapshot.json");
     MockHttpServletResponse response = performCreateSnapshot(snapshotRequest, "_dup_");
     SnapshotSummaryModel summaryModel = validateSnapshotCreated(snapshotRequest, response);
 
     // fetch the snapshot and confirm the metadata matches the request
     SnapshotModel snapshotModel =
-        getTestSnapshot(summaryModel.getId(), snapshotRequest, datasetSummary);
+        SnapshotConnectedTestUtils.getTestSnapshot(
+            mvc, objectMapper, summaryModel.getId(), snapshotRequest, datasetSummary);
     assertNotNull("fetched snapshot successfully after creation", snapshotModel);
 
     // check that the snapshot metadata row is unlocked
@@ -459,429 +448,56 @@ public class SnapshotConnectedTest {
     connectedOperations.getSnapshotExpectError(summaryModel.getId(), HttpStatus.NOT_FOUND);
   }
 
-  @Test
-  public void testExcludeLockedFromSnapshotLookups() throws Exception {
-    // create a snapshot
-    SnapshotSummaryModel snapshotSummary =
-        connectedOperations.createSnapshot(datasetSummary, "snapshot-test-snapshot.json", "_d2_");
-
-    // check that the snapshot metadata row is unlocked
-    String exclusiveLock = snapshotDao.getExclusiveLockState(snapshotSummary.getId());
-    assertNull("snapshot row is unlocked", exclusiveLock);
-
-    // retrieve the snapshot and check that it finds it
-    SnapshotModel snapshotModel = connectedOperations.getSnapshot(snapshotSummary.getId());
-    assertEquals(
-        "Lookup unlocked snapshot succeeds", snapshotSummary.getName(), snapshotModel.getName());
-
-    // enumerate snapshots and check that this snapshot is included in the set
-    EnumerateSnapshotModel enumerateSnapshotModelModel =
-        connectedOperations.enumerateSnapshots(snapshotSummary.getName());
-    List<SnapshotSummaryModel> enumeratedSnapshots = enumerateSnapshotModelModel.getItems();
-    boolean foundSnapshotWithMatchingId = false;
-    for (SnapshotSummaryModel enumeratedSnapshot : enumeratedSnapshots) {
-      if (enumeratedSnapshot.getId().equals(snapshotSummary.getId())) {
-        foundSnapshotWithMatchingId = true;
-        break;
-      }
-    }
-    assertTrue("Unlocked included in enumeration", foundSnapshotWithMatchingId);
-
-    // NO ASSERTS inside the block below where hang is enabled to reduce chance of failing before
-    // disabling the hang
-    // ====================================================
-    // enable hang in DeleteSnapshotPrimaryDataStep
-    configService.setFault(ConfigEnum.SNAPSHOT_DELETE_LOCK_CONFLICT_STOP_FAULT.name(), true);
-
-    // kick off a request to delete the snapshot. this should hang before unlocking the snapshot
-    // object.
-    MvcResult deleteResult =
-        mvc.perform(delete("/api/repository/v1/snapshots/" + snapshotSummary.getId())).andReturn();
-    TimeUnit.SECONDS.sleep(5); // give the flight time to launch
-
-    // note: asserts are below outside the hang block
-    exclusiveLock = snapshotDao.getExclusiveLockState(snapshotSummary.getId());
-
-    // retrieve the snapshot and check that it returns not found
-    // note: asserts are below outside the hang block
-    MvcResult retrieveResult =
-        mvc.perform(get("/api/repository/v1/snapshots/" + snapshotSummary.getId())).andReturn();
-
-    // enumerate snapshots and check that this snapshot is not included in the set
-    // note: asserts are below outside the hang block
-    MvcResult enumerateResult =
-        connectedOperations.enumerateSnapshotsRaw(snapshotSummary.getName());
-
-    // disable hang in DeleteSnapshotPrimaryDataStep
-    configService.setFault(ConfigEnum.SNAPSHOT_DELETE_LOCK_CONFLICT_CONTINUE_FAULT.name(), true);
-    // ====================================================
-
-    // check that the snapshot metadata row has an exclusive lock after kicking off the delete
-    assertNotNull("snapshot row is exclusively locked", exclusiveLock);
-
-    // check that the retrieve snapshot returned not found
-    connectedOperations.handleFailureCase(retrieveResult.getResponse(), HttpStatus.NOT_FOUND);
-
-    // check that the enumerate snapshots returned successfully and that this snapshot is not
-    // included in the set
-    enumerateSnapshotModelModel =
-        connectedOperations.handleSuccessCase(
-            enumerateResult.getResponse(), EnumerateSnapshotModel.class);
-
-    enumeratedSnapshots = enumerateSnapshotModelModel.getItems();
-    foundSnapshotWithMatchingId = false;
-    for (SnapshotSummaryModel enumeratedSnapshot : enumeratedSnapshots) {
-      if (enumeratedSnapshot.getId().equals(snapshotSummary.getId())) {
-        foundSnapshotWithMatchingId = true;
-        break;
-      }
-    }
-    assertFalse("Exclusively locked not included in enumeration", foundSnapshotWithMatchingId);
-
-    // check the response from the delete request
-    MockHttpServletResponse deleteResponse =
-        connectedOperations.validateJobModelAndWait(deleteResult);
-    DeleteResponseModel deleteResponseModel =
-        connectedOperations.handleSuccessCase(deleteResponse, DeleteResponseModel.class);
-    assertEquals(
-        "Snapshot delete returned successfully",
-        DeleteResponseModel.ObjectStateEnum.DELETED,
-        deleteResponseModel.getObjectState());
-
-    // try to fetch the snapshot again and confirm nothing is returned
-    connectedOperations.getSnapshotExpectError(snapshotSummary.getId(), HttpStatus.NOT_FOUND);
-  }
-
-  @Test
-  public void testExcludeLockedFromSnapshotFileLookups() throws Exception {
-    // create a dataset
-    DatasetSummaryModel datasetRefSummary = createTestDataset("simple-with-filerefs-dataset.json");
-
-    // ingest a file
-    URI sourceUri = new URI("gs", "jade-testdata", "/fileloadprofiletest/1KBfile.txt", null, null);
-    String targetFilePath =
-        "/mm/" + Names.randomizeName("testdir") + "/testExcludeLockedFromSnapshotFileLookups.txt";
-    FileLoadModel fileLoadModel =
-        new FileLoadModel()
-            .sourcePath(sourceUri.toString())
-            .description("testExcludeLockedFromSnapshotFileLookups")
-            .mimeType("text/plain")
-            .targetPath(targetFilePath)
-            .profileId(billingProfile.getId());
-    FileModel fileModel =
-        connectedOperations.ingestFileSuccess(datasetRefSummary.getId(), fileLoadModel);
-
-    // generate a JSON file with the fileref
-    String jsonLine = "{\"name\":\"name1\", \"file_ref\":\"" + fileModel.getFileId() + "\"}\n";
-
-    // load a JSON file that contains the table rows to load into the test bucket
-    String jsonFileName = "this-better-pass.json";
-    String dirInCloud =
-        "scratch/testExcludeLockedFromSnapshotFileLookups/" + UUID.randomUUID().toString();
-    BlobInfo ingestTableBlob =
-        BlobInfo.newBuilder(testConfig.getIngestbucket(), dirInCloud + "/" + jsonFileName).build();
-    Storage storage = StorageOptions.getDefaultInstance().getService();
-    storage.create(ingestTableBlob, jsonLine.getBytes(StandardCharsets.UTF_8));
-
-    // make sure the JSON file gets cleaned up on test teardown
-    connectedOperations.addScratchFile(dirInCloud + "/" + jsonFileName);
-
-    // ingest the tabular data from the JSON file we just generated
-    String gsPath = "gs://" + testConfig.getIngestbucket() + "/" + dirInCloud + "/" + jsonFileName;
-    IngestRequestModel ingestRequest1 =
-        new IngestRequestModel()
-            .format(IngestRequestModel.FormatEnum.JSON)
-            .table("tableA")
-            .path(gsPath);
-    connectedOperations.ingestTableSuccess(datasetRefSummary.getId(), ingestRequest1);
-
-    // create a snapshot
-    SnapshotSummaryModel snapshotSummary =
-        connectedOperations.createSnapshot(
-            datasetRefSummary, "simple-with-filerefs-snapshot.json", "");
-
-    // check that the snapshot metadata row is unlocked
-    String exclusiveLock = snapshotDao.getExclusiveLockState(snapshotSummary.getId());
-    assertNull("snapshot row is unlocked", exclusiveLock);
-
-    /*
-     * WARNING: if making any changes to this test make sure to notify the #dsp-batch channel! Describe the change
-     * and any consequences downstream to DRS clients.
-     */
-    String fileUri = getFileRefIdFromSnapshot(snapshotSummary);
-    DrsId drsId = drsIdService.fromUri(fileUri);
-    DRSObject drsObject = connectedOperations.drsGetObjectSuccess(drsId.toDrsObjectId(), false);
-    String filePath = drsObject.getAliases().get(0);
-
-    // lookup the snapshot file by DRS id, make sure it's returned (lookupSnapshotFileSuccess will
-    // already check)
-    FileModel fsObjById =
-        connectedOperations.lookupSnapshotFileSuccess(
-            snapshotSummary.getId(), drsId.getFsObjectId());
-    assertEquals(
-        "Retrieve snapshot file by DRS id matches desc",
-        fsObjById.getDescription(),
-        fileLoadModel.getDescription());
-
-    // lookup the snapshot file by DRS path and check that it's found
-    FileModel fsObjByPath =
-        connectedOperations.lookupSnapshotFileByPathSuccess(snapshotSummary.getId(), filePath, 0);
-    assertEquals(
-        "Retrieve snapshot file by path matches desc",
-        fsObjByPath.getDescription(),
-        fileLoadModel.getDescription());
-    assertThat(
-        "Retrieve snapshot file objects match", fsObjById, CoreMatchers.equalTo(fsObjByPath));
-
-    // now the snapshot exists....let's get it locked!
-
-    // NO ASSERTS inside the block below where hang is enabled to reduce chance of failing before
-    // disabling the hang
-    // ====================================================
-    // enable hang in DeleteSnapshotPrimaryDataStep
-    configService.setFault(ConfigEnum.SNAPSHOT_DELETE_LOCK_CONFLICT_STOP_FAULT.name(), true);
-
-    // kick off a request to delete the snapshot. this should hang before unlocking the snapshot
-    // object.
-    // note: asserts are below outside the hang block
-    MvcResult deleteResult =
-        mvc.perform(delete("/api/repository/v1/snapshots/" + snapshotSummary.getId())).andReturn();
-    TimeUnit.SECONDS.sleep(5); // give the flight time to launch and get to the hang
-
-    // check that the snapshot metadata row has an exclusive lock
-    exclusiveLock = snapshotDao.getExclusiveLockState(snapshotSummary.getId());
-
-    // lookup the snapshot file by id and check that it's NOT found
-    MockHttpServletResponse failedGetSnapshotByIdResponse =
-        connectedOperations.lookupSnapshotFileRaw(snapshotSummary.getId(), drsId.getFsObjectId());
-
-    // lookup the snapshot file by path and check that it's NOT found
-    MockHttpServletResponse failedGetSnapshotByPathResponse =
-        connectedOperations.lookupSnapshotFileByPathRaw(snapshotSummary.getId(), filePath, 0);
-
-    // disable hang in DeleteSnapshotPrimaryDataStep
-    configService.setFault(ConfigEnum.SNAPSHOT_DELETE_LOCK_CONFLICT_CONTINUE_FAULT.name(), true);
-    // ====================================================
-
-    // check that the snapshot metadata row has an exclusive lock after kicking off the delete
-    assertNotNull("snapshot row is exclusively locked", exclusiveLock);
-
-    assertEquals(
-        "Snapshot file NOT found by DRS id lookup",
-        HttpStatus.NOT_FOUND,
-        HttpStatus.valueOf(failedGetSnapshotByIdResponse.getStatus()));
-
-    assertEquals(
-        "Snapshot file NOT found by path lookup",
-        HttpStatus.NOT_FOUND,
-        HttpStatus.valueOf(failedGetSnapshotByPathResponse.getStatus()));
-
-    // check the response from the snapshot delete request
-    MockHttpServletResponse deleteResponse =
-        connectedOperations.validateJobModelAndWait(deleteResult);
-    DeleteResponseModel deleteResponseModel =
-        connectedOperations.handleSuccessCase(deleteResponse, DeleteResponseModel.class);
-    assertEquals(
-        "Snapshot delete returned successfully",
-        DeleteResponseModel.ObjectStateEnum.DELETED,
-        deleteResponseModel.getObjectState());
-
-    // delete the dataset and check that it succeeds
-    connectedOperations.deleteTestDatasetAndCleanup(datasetRefSummary.getId());
-
-    // remove the file from the connectedoperation bookkeeping list
-    connectedOperations.removeFile(datasetRefSummary.getId(), fileModel.getFileId());
-
-    // try to fetch the snapshot again and confirm nothing is returned
-    connectedOperations.getSnapshotExpectError(snapshotSummary.getId(), HttpStatus.NOT_FOUND);
-
-    // try to fetch the dataset again and confirm nothing is returned
-    connectedOperations.getDatasetExpectError(datasetRefSummary.getId(), HttpStatus.NOT_FOUND);
-  }
-
-  @Test
-  public void testSnapshotArrayFileRefLookups() throws Exception {
-    // create a dataset
-    DatasetSummaryModel datasetRefSummary =
-        createTestDataset("simple-with-array-filerefs-dataset.json");
-
-    // ingest a file
-    URI sourceUri = new URI("gs", "jade-testdata", "/fileloadprofiletest/1KBfile.txt", null, null);
-    String targetFilePath =
-        "/mm/" + Names.randomizeName("testdir") + "/testExcludeLockedFromSnapshotFileLookups.txt";
-    UUID descriptionRandomUUID = UUID.randomUUID();
-    FileLoadModel fileLoadModel =
-        new FileLoadModel()
-            .sourcePath(sourceUri.toString())
-            .description(descriptionRandomUUID.toString())
-            .mimeType("text/plain")
-            .targetPath(targetFilePath)
-            .profileId(billingProfile.getId());
-    FileModel fileModel =
-        connectedOperations.ingestFileSuccess(datasetRefSummary.getId(), fileLoadModel);
-
-    // generate a JSON file with the fileref
-    String jsonLine = "{\"name\":\"name1\", \"file_ref\":[\"" + fileModel.getFileId() + "\"]}\n";
-
-    // load a JSON file that contains the table rows to load into the test bucket
-    String jsonFileName = "this-array-better-pass.json";
-    String dirInCloud =
-        "scratch/testExcludeLockedFromSnapshotFileLookups/" + UUID.randomUUID().toString();
-    BlobInfo ingestTableBlob =
-        BlobInfo.newBuilder(testConfig.getIngestbucket(), dirInCloud + "/" + jsonFileName).build();
-    Storage storage = StorageOptions.getDefaultInstance().getService();
-    storage.create(ingestTableBlob, jsonLine.getBytes(StandardCharsets.UTF_8));
-
-    // make sure the JSON file gets cleaned up on test teardown
-    connectedOperations.addScratchFile(dirInCloud + "/" + jsonFileName);
-
-    // ingest the tabular data from the JSON file we just generated
-    String gsPath = "gs://" + testConfig.getIngestbucket() + "/" + dirInCloud + "/" + jsonFileName;
-    IngestRequestModel ingestRequest1 =
-        new IngestRequestModel()
-            .format(IngestRequestModel.FormatEnum.JSON)
-            .table("tableA")
-            .path(gsPath);
-    connectedOperations.ingestTableSuccess(datasetRefSummary.getId(), ingestRequest1);
-
-    // create a snapshot
-    SnapshotSummaryModel snapshotSummary =
-        connectedOperations.createSnapshot(
-            datasetRefSummary, "simple-with-filerefs-snapshot.json", "");
-
-    // check that the snapshot metadata row is unlocked
-    String exclusiveLock = snapshotDao.getExclusiveLockState(snapshotSummary.getId());
-    assertNull("snapshot row is unlocked", exclusiveLock);
-
-    /*
-     * WARNING: if making any changes to this test make sure to notify the #dsp-batch channel! Describe the change
-     * and any consequences downstream to DRS clients.
-     */
-    String fileUri = getFileRefIdFromSnapshot(snapshotSummary);
-    DrsId drsId = drsIdService.fromUri(fileUri);
-    DRSObject drsObject = connectedOperations.drsGetObjectSuccess(drsId.toDrsObjectId(), false);
-    String filePath = drsObject.getAliases().get(0);
-
-    // lookup the snapshot file by DRS id, make sure it's returned (lookupSnapshotFileSuccess will
-    // already check)
-    FileModel fsObjById =
-        connectedOperations.lookupSnapshotFileSuccess(
-            snapshotSummary.getId(), drsId.getFsObjectId());
-    assertEquals(
-        "Retrieve snapshot file by DRS id matches uuid in desc",
-        fsObjById.getDescription(),
-        fileLoadModel.getDescription());
-
-    // lookup the snapshot file by DRS path and check that it's found
-    FileModel fsObjByPath =
-        connectedOperations.lookupSnapshotFileByPathSuccess(snapshotSummary.getId(), filePath, 0);
-    assertEquals(
-        "Retrieve snapshot file by path matches uuid in desc",
-        fsObjByPath.getDescription(),
-        fileLoadModel.getDescription());
-    assertThat(
-        "Retrieve snapshot file objects match", fsObjById, CoreMatchers.equalTo(fsObjByPath));
-  }
-
   private DatasetSummaryModel setupMinimalDataset() throws Exception {
-    DatasetSummaryModel datasetMinimalSummary = createTestDataset("dataset-minimal.json");
-    loadCsvData(datasetMinimalSummary.getId(), "participant", "dataset-minimal-participant.csv");
-    loadCsvData(datasetMinimalSummary.getId(), "sample", "dataset-minimal-sample.csv");
+    DatasetSummaryModel datasetMinimalSummary =
+        SnapshotConnectedTestUtils.createTestDataset(
+            connectedOperations, billingProfile, "dataset-minimal.json");
+    SnapshotConnectedTestUtils.loadCsvData(
+        connectedOperations,
+        jsonLoader,
+        storage,
+        testConfig.getIngestbucket(),
+        datasetMinimalSummary.getId(),
+        "participant",
+        "dataset-minimal-participant.csv");
+    SnapshotConnectedTestUtils.loadCsvData(
+        connectedOperations,
+        jsonLoader,
+        storage,
+        testConfig.getIngestbucket(),
+        datasetMinimalSummary.getId(),
+        "sample",
+        "dataset-minimal-sample.csv");
     return datasetMinimalSummary;
   }
 
   private DatasetSummaryModel setupArrayStructDataset() throws Exception {
-    DatasetSummaryModel datasetArraySummary = createTestDataset("dataset-array-struct.json");
+    DatasetSummaryModel datasetArraySummary =
+        SnapshotConnectedTestUtils.createTestDataset(
+            connectedOperations, billingProfile, "dataset-array-struct.json");
     loadJsonData(
         datasetArraySummary.getId(), "participant", "dataset-array-struct-participant.json");
     loadJsonData(datasetArraySummary.getId(), "sample", "dataset-array-struct-sample.json");
     return datasetArraySummary;
   }
 
-  // create a dataset to create snapshots in and return its id
-  private DatasetSummaryModel createTestDataset(String resourcePath) throws Exception {
-    return connectedOperations.createDataset(billingProfile, resourcePath);
-  }
-
-  private void loadCsvData(UUID datasetId, String tableName, String resourcePath) throws Exception {
-    loadData(datasetId, tableName, resourcePath, IngestRequestModel.FormatEnum.CSV);
-  }
-
   private void loadJsonData(UUID datasetId, String tableName, String resourcePath)
       throws Exception {
-    loadData(datasetId, tableName, resourcePath, IngestRequestModel.FormatEnum.JSON);
-  }
-
-  private void loadData(
-      UUID datasetId, String tableName, String resourcePath, IngestRequestModel.FormatEnum format)
-      throws Exception {
-
-    String bucket = testConfig.getIngestbucket();
-    BlobInfo stagingBlob =
-        BlobInfo.newBuilder(bucket, UUID.randomUUID() + "-" + resourcePath).build();
-    byte[] data = IOUtils.toByteArray(jsonLoader.getClassLoader().getResource(resourcePath));
-
-    IngestRequestModel ingestRequest =
-        new IngestRequestModel()
-            .table(tableName)
-            .format(format)
-            .path("gs://" + stagingBlob.getBucket() + "/" + stagingBlob.getName());
-
-    if (format.equals(IngestRequestModel.FormatEnum.CSV)) {
-      ingestRequest.csvSkipLeadingRows(1);
-      ingestRequest.csvGenerateRowIds(false);
-    }
-
-    try {
-      storage.create(stagingBlob, data);
-      connectedOperations.ingestTableSuccess(datasetId, ingestRequest);
-    } finally {
-      storage.delete(stagingBlob.getBlobId());
-    }
-  }
-
-  private SnapshotRequestModel makeSnapshotTestRequest(
-      DatasetSummaryModel datasetSummaryModel, String resourcePath) throws Exception {
-    SnapshotRequestModel snapshotRequest =
-        jsonLoader.loadObject(resourcePath, SnapshotRequestModel.class);
-    SnapshotRequestContentsModel content = snapshotRequest.getContents().get(0);
-    // TODO SingleDatasetSnapshot
-    String newDatasetName = datasetSummaryModel.getName();
-    String origDatasetName = content.getDatasetName();
-    // swap in the correct dataset name (with the id at the end)
-    content.setDatasetName(newDatasetName);
-    snapshotRequest.profileId(datasetSummaryModel.getDefaultProfileId());
-    if (content.getMode().equals(SnapshotRequestContentsModel.ModeEnum.BYQUERY)) {
-      // if its by query, also set swap in the correct dataset name in the query
-      String query = content.getQuerySpec().getQuery();
-      content.getQuerySpec().setQuery(query.replace(origDatasetName, newDatasetName));
-    }
-    return snapshotRequest;
-  }
-
-  private MvcResult launchCreateSnapshot(SnapshotRequestModel snapshotRequest, String infix)
-      throws Exception {
-    if (infix != null) {
-      String snapshotName = Names.randomizeNameInfix(snapshotRequest.getName(), infix);
-      snapshotRequest.setName(snapshotName);
-    }
-
-    String jsonRequest = TestUtils.mapToJson(snapshotRequest);
-
-    return mvc.perform(
-            post("/api/repository/v1/snapshots")
-                .contentType(MediaType.APPLICATION_JSON)
-                .content(jsonRequest))
-        // TODO: swagger field validation errors do not set content type; they log and return
-        // nothing
-        .andExpect(content().contentType(MediaType.APPLICATION_JSON))
-        .andReturn();
+    SnapshotConnectedTestUtils.loadData(
+        connectedOperations,
+        jsonLoader,
+        storage,
+        testConfig.getIngestbucket(),
+        datasetId,
+        tableName,
+        resourcePath,
+        IngestRequestModel.FormatEnum.JSON);
   }
 
   private MockHttpServletResponse performCreateSnapshot(
       SnapshotRequestModel snapshotRequest, String infix) throws Exception {
     snapshotOriginalName = snapshotRequest.getName();
-    MvcResult result = launchCreateSnapshot(snapshotRequest, infix);
+    MvcResult result = SnapshotConnectedTestUtils.launchCreateSnapshot(mvc, snapshotRequest, infix);
     MockHttpServletResponse response = connectedOperations.validateJobModelAndWait(result);
     return response;
   }
@@ -924,32 +540,6 @@ public class SnapshotConnectedTest {
     return summary;
   }
 
-  private SnapshotModel getTestSnapshot(
-      UUID id, SnapshotRequestModel snapshotRequest, DatasetSummaryModel datasetSummary)
-      throws Exception {
-    MvcResult result =
-        mvc.perform(get("/api/repository/v1/snapshots/" + id))
-            .andExpect(content().contentType(MediaType.APPLICATION_JSON_VALUE))
-            .andReturn();
-
-    MockHttpServletResponse response = result.getResponse();
-    SnapshotModel snapshotModel =
-        objectMapper.readValue(response.getContentAsString(), SnapshotModel.class);
-
-    assertThat(snapshotModel.getDescription(), equalTo(snapshotRequest.getDescription()));
-    assertThat(snapshotModel.getName(), startsWith(snapshotRequest.getName()));
-
-    assertThat("source array has one element", snapshotModel.getSource().size(), equalTo(1));
-    SnapshotSourceModel sourceModel = snapshotModel.getSource().get(0);
-
-    assertThat(
-        "snapshot dataset summary is the same as from dataset",
-        sourceModel.getDataset(),
-        equalTo(datasetSummary));
-
-    return snapshotModel;
-  }
-
   private static final String queryForCountTemplate =
       "SELECT COUNT(*) FROM `<project>.<snapshot>.<table>`";
 
@@ -970,51 +560,5 @@ public class SnapshotConnectedTest {
     FieldValueList row = result.iterateAll().iterator().next();
     FieldValue countValue = row.get(0);
     return countValue.getLongValue();
-  }
-
-  private static final String queryForRefIdTemplate =
-      "SELECT file_ref FROM `<project>.<snapshot>.<table>` WHERE file_ref IS NOT NULL";
-
-  // Technically a helper method, but so specific to testExcludeLockedFromSnapshotFileLookups,
-  // likely not re-useable
-  private String getFileRefIdFromSnapshot(SnapshotSummaryModel snapshotSummary)
-      throws InterruptedException {
-    Snapshot snapshot = snapshotDao.retrieveSnapshotByName(snapshotSummary.getName());
-    BigQueryProject bigQueryProject =
-        BigQueryProject.get(snapshot.getProjectResource().getGoogleProjectId());
-    BigQuery bigQuery = bigQueryProject.getBigQuery();
-
-    ST sqlTemplate = new ST(queryForRefIdTemplate);
-    sqlTemplate.add("project", snapshot.getProjectResource().getGoogleProjectId());
-    sqlTemplate.add("snapshot", snapshot.getName());
-    sqlTemplate.add("table", "tableA");
-
-    QueryJobConfiguration queryConfig =
-        QueryJobConfiguration.newBuilder(sqlTemplate.render()).build();
-    TableResult result = bigQuery.query(queryConfig);
-    FieldValueList row = result.iterateAll().iterator().next();
-    FieldValue idValue = row.get(0);
-    if (idValue.getAttribute().name().equals("REPEATED")) {
-      return idValue.getRepeatedValue().stream()
-          .map(FieldValue::getStringValue)
-          .collect(Collectors.toList())
-          .get(0);
-    } else {
-      return idValue.getStringValue();
-    }
-  }
-
-  private void snapshotHappyPathTestingHelper(String path) throws Exception {
-    SnapshotRequestModel snapshotRequest = makeSnapshotTestRequest(datasetSummary, path);
-    MockHttpServletResponse response = performCreateSnapshot(snapshotRequest, "_thp_");
-    SnapshotSummaryModel summaryModel = validateSnapshotCreated(snapshotRequest, response);
-
-    SnapshotModel snapshotModel =
-        getTestSnapshot(summaryModel.getId(), snapshotRequest, datasetSummary);
-
-    connectedOperations.deleteTestSnapshot(snapshotModel.getId());
-    // Duplicate delete should work
-    connectedOperations.deleteTestSnapshot(snapshotModel.getId());
-    connectedOperations.getSnapshotExpectError(snapshotModel.getId(), HttpStatus.NOT_FOUND);
   }
 }
