+++ src/test/java/bio/terra/service/tabulardata/google/BigQueryPdaoTest.java
@@ -5,12 +5,10 @@ import static bio.terra.common.PdaoConstant.PDAO_LOAD_HISTORY_TABLE;
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.containsInAnyOrder;
-import static org.hamcrest.Matchers.empty;
 import static org.hamcrest.Matchers.equalTo;
 import static org.junit.Assert.assertEquals;
 
 import bio.terra.app.configuration.ConnectedTestConfiguration;
-import bio.terra.app.model.GoogleCloudResource;
 import bio.terra.app.model.GoogleRegion;
 import bio.terra.buffer.model.ResourceInfo;
 import bio.terra.common.CloudPlatformWrapper;
@@ -42,9 +40,7 @@ import bio.terra.service.snapshot.Snapshot;
 import bio.terra.service.snapshot.SnapshotDao;
 import bio.terra.service.tabulardata.exception.BadExternalFileException;
 import com.google.cloud.bigquery.BigQuery;
-import com.google.cloud.bigquery.JobInfo;
 import com.google.cloud.bigquery.QueryJobConfiguration;
-import com.google.cloud.bigquery.TableId;
 import com.google.cloud.bigquery.TableResult;
 import com.google.cloud.storage.BlobInfo;
 import com.google.cloud.storage.Storage;
@@ -165,179 +161,6 @@ public class BigQueryPdaoTest {
       assertThatDatasetAndTablesShouldExist(dataset, false);
     } finally {
       datasetDao.delete(dataset.getId());
-    }
-  }
-
-  @Test
-  public void datasetTest() throws Exception {
-    Dataset defaultDataset = readDataset("ingest-test-dataset.json");
-    Dataset eastDataset = readDataset("ingest-test-dataset-east.json");
-    List<Object[]> testCases =
-        Arrays.asList(
-            new Object[] {
-              defaultDataset,
-              testConfig.getIngestbucket(),
-              "BiqQuery datasets are instantiated in us-central1 by default."
-            },
-            new Object[] {
-              eastDataset,
-              testConfig.getNonDefaultRegionIngestBucket(),
-              "BiqQuery datasets can be set to the non-default region."
-            });
-
-    for (Object[] tuple : testCases) {
-      String targetPath = "scratch/file" + UUID.randomUUID().toString() + "/";
-      Dataset dataset = (Dataset) tuple[0];
-      String bucket = (String) tuple[1];
-      String regionMessage = (String) tuple[2];
-
-      String region =
-          dataset
-              .getDatasetSummary()
-              .getStorageResourceRegion(GoogleCloudResource.BIGQUERY)
-              .toString();
-
-      connectedOperations.addDataset(dataset.getId());
-
-      // Stage tabular data for ingest.
-      BlobInfo participantBlob =
-          BlobInfo.newBuilder(bucket, targetPath + "ingest-test-participant.json").build();
-      BlobInfo sampleBlob =
-          BlobInfo.newBuilder(bucket, targetPath + "ingest-test-sample.json").build();
-      BlobInfo fileBlob = BlobInfo.newBuilder(bucket, targetPath + "ingest-test-file.json").build();
-
-      BlobInfo missingPkBlob =
-          BlobInfo.newBuilder(bucket, targetPath + "ingest-test-sample-no-id.json").build();
-      BlobInfo nullPkBlob =
-          BlobInfo.newBuilder(bucket, targetPath + "ingest-test-sample-null-id.json").build();
-
-      try {
-        bigQueryPdao.createDataset(dataset);
-
-        com.google.cloud.bigquery.Dataset bqDataset = bigQueryDataset(dataset);
-        assertThat(regionMessage, bqDataset.getLocation(), equalTo(region));
-
-        storage.create(participantBlob, readFile("ingest-test-participant.json"));
-        storage.create(sampleBlob, readFile("ingest-test-sample.json"));
-        storage.create(fileBlob, readFile("ingest-test-file.json"));
-        storage.create(missingPkBlob, readFile("ingest-test-sample-no-id.json"));
-        storage.create(nullPkBlob, readFile("ingest-test-sample-null-id.json"));
-
-        // Ingest staged data into the new dataset.
-        IngestRequestModel ingestRequest =
-            new IngestRequestModel().format(IngestRequestModel.FormatEnum.JSON);
-
-        UUID datasetId = dataset.getId();
-        connectedOperations.ingestTableSuccess(
-            datasetId, ingestRequest.table("participant").path(gsPath(participantBlob)));
-        connectedOperations.ingestTableSuccess(
-            datasetId, ingestRequest.table("sample").path(gsPath(sampleBlob)));
-        connectedOperations.ingestTableSuccess(
-            datasetId, ingestRequest.table("file").path(gsPath(fileBlob)));
-
-        // Check primary key non-nullability is enforced.
-        connectedOperations.ingestTableFailure(
-            datasetId, ingestRequest.table("sample").path(gsPath(missingPkBlob)));
-        connectedOperations.ingestTableFailure(
-            datasetId, ingestRequest.table("sample").path(gsPath(nullPkBlob)));
-
-        // Create a snapshot!
-        DatasetSummaryModel datasetSummaryModel =
-            DatasetJsonConversion.datasetSummaryModelFromDatasetSummary(
-                dataset.getDatasetSummary());
-        SnapshotSummaryModel snapshotSummary =
-            connectedOperations.createSnapshot(
-                datasetSummaryModel, "ingest-test-snapshot.json", "");
-        SnapshotModel snapshot = connectedOperations.getSnapshot(snapshotSummary.getId());
-
-        com.google.cloud.bigquery.Dataset bqSnapshotDataset = bigQuerySnapshot(snapshot);
-
-        assertThat(
-            String.format(
-                "Snapshot for dataset in region %s should also also be in region %s",
-                region, region),
-            region,
-            equalTo(bqSnapshotDataset.getLocation()));
-
-        BigQueryProject bigQueryProject =
-            TestUtils.bigQueryProjectForDatasetName(datasetDao, dataset.getName());
-        BigQueryProject bigQuerySnapshotProject =
-            TestUtils.bigQueryProjectForSnapshotName(snapshotDao, snapshot.getName());
-        assertThat(snapshot.getTables().size(), is(equalTo(3)));
-        List<String> participantIds =
-            queryForIds(snapshot.getName(), "participant", bigQuerySnapshotProject);
-        List<String> sampleIds = queryForIds(snapshot.getName(), "sample", bigQuerySnapshotProject);
-        List<String> fileIds = queryForIds(snapshot.getName(), "file", bigQuerySnapshotProject);
-
-        assertThat(
-            participantIds,
-            containsInAnyOrder(
-                "participant_1",
-                "participant_2",
-                "participant_3",
-                "participant_4",
-                "participant_5"));
-        assertThat(sampleIds, containsInAnyOrder("sample1", "sample2", "sample5"));
-        assertThat(fileIds, is(equalTo(Collections.singletonList("file1"))));
-
-        // Simulate soft-deleting some rows.
-        // TODO: Replace this with a call to the soft-delete API once it exists?
-        softDeleteRows(
-            bigQueryProject,
-            bigQueryPdao.prefixName(dataset.getName()),
-            getTable(dataset, "participant"),
-            Arrays.asList("participant_3", "participant_4"));
-        softDeleteRows(
-            bigQueryProject,
-            bigQueryPdao.prefixName(dataset.getName()),
-            getTable(dataset, "sample"),
-            Collections.singletonList("sample5"));
-        softDeleteRows(
-            bigQueryProject,
-            bigQueryPdao.prefixName(dataset.getName()),
-            getTable(dataset, "file"),
-            Collections.singletonList("file1"));
-
-        // Create another snapshot.
-        snapshotSummary =
-            connectedOperations.createSnapshot(
-                datasetSummaryModel, "ingest-test-snapshot.json", "");
-        SnapshotModel snapshot2 = connectedOperations.getSnapshot(snapshotSummary.getId());
-        assertThat(snapshot2.getTables().size(), is(equalTo(3)));
-
-        BigQueryProject bigQuerySnapshotProject2 =
-            TestUtils.bigQueryProjectForSnapshotName(snapshotDao, snapshot2.getName());
-
-        participantIds = queryForIds(snapshot2.getName(), "participant", bigQuerySnapshotProject2);
-        sampleIds = queryForIds(snapshot2.getName(), "sample", bigQuerySnapshotProject2);
-        fileIds = queryForIds(snapshot2.getName(), "file", bigQuerySnapshotProject2);
-        assertThat(
-            participantIds, containsInAnyOrder("participant_1", "participant_2", "participant_5"));
-        assertThat(sampleIds, containsInAnyOrder("sample1", "sample2"));
-        assertThat(fileIds, is(empty()));
-
-        // Make sure the old snapshot wasn't changed.
-        participantIds = queryForIds(snapshot.getName(), "participant", bigQuerySnapshotProject);
-        sampleIds = queryForIds(snapshot.getName(), "sample", bigQuerySnapshotProject);
-        fileIds = queryForIds(snapshot.getName(), "file", bigQuerySnapshotProject);
-        assertThat(
-            participantIds,
-            containsInAnyOrder(
-                "participant_1",
-                "participant_2",
-                "participant_3",
-                "participant_4",
-                "participant_5"));
-        assertThat(sampleIds, containsInAnyOrder("sample1", "sample2", "sample5"));
-        assertThat(fileIds, is(equalTo(Collections.singletonList("file1"))));
-      } finally {
-        storage.delete(
-            participantBlob.getBlobId(),
-            sampleBlob.getBlobId(),
-            fileBlob.getBlobId(),
-            missingPkBlob.getBlobId(),
-            nullPkBlob.getBlobId());
-      }
     }
   }
 
@@ -564,13 +387,7 @@ public class BigQueryPdaoTest {
     return values;
   }
 
-  public com.google.cloud.bigquery.Dataset bigQueryDataset(Dataset dataset) {
-    return BigQueryProject.from(dataset)
-        .getBigQuery()
-        .getDataset(bigQueryPdao.prefixName(dataset.getName()));
-  }
-
-  public com.google.cloud.bigquery.Dataset bigQuerySnapshot(SnapshotModel snapshot) {
+  static com.google.cloud.bigquery.Dataset bigQuerySnapshot(SnapshotModel snapshot) {
     return BigQueryProject.from(snapshot).getBigQuery().getDataset(snapshot.getName());
   }
 
@@ -578,40 +395,15 @@ public class BigQueryPdaoTest {
     return IOUtils.toByteArray(getClass().getClassLoader().getResource(fileName));
   }
 
-  private String gsPath(BlobInfo blob) {
+  static String gsPath(BlobInfo blob) {
     return "gs://" + blob.getBucket() + "/" + blob.getName();
-  }
-
-  private static final String queryAllRowIdsTemplate =
-      "SELECT "
-          + PdaoConstant.PDAO_ROW_ID_COLUMN
-          + " FROM `<project>.<dataset>.<table>` "
-          + "WHERE id IN UNNEST([<ids:{id|'<id>'}; separator=\",\">])";
-
-  private void softDeleteRows(
-      BigQueryProject bq, String datasetName, DatasetTable table, List<String> ids)
-      throws Exception {
-
-    ST sqlTemplate = new ST(queryAllRowIdsTemplate);
-    sqlTemplate.add("project", bq.getProjectId());
-    sqlTemplate.add("dataset", datasetName);
-    sqlTemplate.add("table", table.getRawTableName());
-    sqlTemplate.add("ids", ids);
-
-    QueryJobConfiguration queryConfig =
-        QueryJobConfiguration.newBuilder(sqlTemplate.render())
-            .setDestinationTable(TableId.of(datasetName, table.getSoftDeleteTableName()))
-            .setWriteDisposition(JobInfo.WriteDisposition.WRITE_APPEND)
-            .build();
-
-    bq.getBigQuery().query(queryConfig);
   }
 
   private static final String queryForIdsTemplate =
       "SELECT id FROM `<project>.<snapshot>.<table>` ORDER BY id";
 
   // Get the count of rows in a table or view
-  private List<String> queryForIds(
+  static List<String> queryForIds(
       String snapshotName, String tableName, BigQueryProject bigQueryProject) throws Exception {
     String bigQueryProjectId = bigQueryProject.getProjectId();
     BigQuery bigQuery = bigQueryProject.getBigQuery();
@@ -649,7 +441,7 @@ public class BigQueryPdaoTest {
           + PdaoConstant.PDAO_INGEST_DATE_COLUMN_ALIAS
           + " FROM `<project>.<dataset>.<table>`";
 
-  private String makeDatasetName() {
+  static String makeDatasetName() {
     return "pdaotest" + StringUtils.remove(UUID.randomUUID().toString(), '-');
   }
 
@@ -687,7 +479,7 @@ public class BigQueryPdaoTest {
     return dataset;
   }
 
-  private DatasetTable getTable(Dataset dataset, String name) {
+  static DatasetTable getTable(Dataset dataset, String name) {
     return dataset
         .getTableByName(name)
         .orElseThrow(() -> new IllegalStateException("Expected table " + name + " not found!"));
